refactor: split journal/utils.ts into focused modules
399-line file -> templates.ts, paths.ts, editor.ts, fs.ts with
barrel index.ts. Each file has a single responsibility.
