+++ src/commands/journal/meeting.ts
@@ -12,7 +12,7 @@ import {
   ensureTemplatesExist,
   generateJournalFileInfoByType,
   openInEditor,
-} from "../../lib/journal/utils.js";
+} from "../../lib/journal/index.js";
 
 /**
  * Create or open meeting notes file
