+++ src/commands/journal/note.ts
@@ -12,7 +12,7 @@ import {
   ensureTemplatesExist,
   generateJournalFileInfoByType,
   openInEditor,
-} from "../../lib/journal/utils.js";
+} from "../../lib/journal/index.js";
 
 /**
  * Create or open general-purpose note file
