+++ src/lib/journal/index.ts
@@ -1,0 +1,11 @@
+export { ensureDirectoryExists } from "./fs.js";
+export { openInEditor } from "./editor.js";
+export {
+  loadTemplate,
+  ensureTemplatesExist,
+  createJournalContent,
+  createMeetingContent,
+  createNoteContent,
+} from "./templates.js";
+export { resolvePathTemplate, generateJournalFileInfoByType } from "./paths.js";
+export type { GenerateJournalFileResult, GenerateJournalFileParams } from "./paths.js";
